HiMolformer/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
absl-py=2.0.0=pypi_0
aiohttp=3.8.6=pypi_0
aiosignal=1.3.1=pypi_0
anyio=3.7.1=pypi_0
argon2-cffi=23.1.0=pypi_0
argon2-cffi-bindings=21.2.0=pypi_0
asttokens=2.4.1=pypi_0
async-timeout=4.0.3=pypi_0
attrs=23.1.0=pypi_0
babel=2.13.1=pypi_0
backcall=0.2.0=pypi_0
beautifulsoup4=4.12.2=pypi_0
bertviz=1.4.0=pypi_0
bleach=6.1.0=pypi_0
boost=1.74.0=py39h5472131_5
boost-cpp=1.74.0=h75c5d50_8
boto3=1.28.73=pypi_0
botocore=1.31.73=pypi_0
bzip2=1.0.8=h7f98852_4
ca-certificates=2023.7.22=hbcca054_0
cachetools=5.3.2=pypi_0
cairo=1.16.0=hb05425b_5
certifi=2022.12.7=pypi_0
cffi=1.16.0=pypi_0
charset-normalizer=2.1.1=pypi_0
click=8.1.7=pypi_0
comm=0.1.4=pypi_0
cycler=0.12.1=pyhd8ed1ab_0
datasets=1.6.2=pypi_0
debugpy=1.8.0=pypi_0
decorator=5.1.1=pypi_0
defusedxml=0.7.1=pypi_0
dill=0.3.7=pypi_0
entrypoints=0.4=pypi_0
exceptiongroup=1.1.3=pypi_0
executing=2.0.0=pypi_0
expat=2.5.0=hcb278e6_1
fastjsonschema=2.18.1=pypi_0
filelock=3.9.0=pypi_0
fontconfig=2.14.1=h52c9d5c_1
freetype=2.10.4=hca18f0e_2
frozenlist=1.4.0=pypi_0
fsspec=2023.4.0=pypi_0
future=0.18.3=pypi_0
giflib=5.2.1=h0b41bf4_3
glib=2.69.1=h4ff587b_1
google-auth=2.23.3=pypi_0
google-auth-oauthlib=1.1.0=pypi_0
greenlet=3.0.1=py39h3d6467e_0
grpcio=1.59.0=pypi_0
huggingface-hub=0.0.8=pypi_0
icu=70.1=h27087fc_0
idna=3.4=pypi_0
importlib-metadata=6.8.0=pypi_0
ipykernel=6.26.0=pypi_0
ipython=8.16.1=pypi_0
ipython-genutils=0.2.0=pypi_0
ipywidgets=7.7.0=pypi_0
jedi=0.19.1=pypi_0
jinja2=3.1.2=pypi_0
jmespath=1.0.1=pypi_0
joblib=1.3.2=pypi_0
jpeg=9e=h0b41bf4_3
json5=0.9.14=pypi_0
jsonschema=4.19.1=pypi_0
jsonschema-specifications=2023.7.1=pypi_0
jupyter-client=7.4.9=pypi_0
jupyter-core=5.4.0=pypi_0
jupyter-server=1.24.0=pypi_0
jupyterlab=3.4.0=pypi_0
jupyterlab-pygments=0.2.2=pypi_0
jupyterlab-server=2.25.0=pypi_0
jupyterlab-widgets=3.0.9=pypi_0
kiwisolver=1.4.5=py39h7633fee_1
lcms2=2.15=hfd0df8a_0
ld_impl_linux-64=2.40=h41732ed_0
lerc=3.0=h295c915_0
libblas=3.9.0=19_linux64_openblas
libcblas=3.9.0=19_linux64_openblas
libdeflate=1.17=h5eee18b_1
libexpat=2.5.0=hcb278e6_1
libffi=3.3=h58526e2_2
libgcc-ng=13.2.0=h807b86a_2
libgfortran-ng=13.2.0=h69a702a_2
libgfortran5=13.2.0=ha4646dd_2
libgomp=13.2.0=h807b86a_2
libiconv=1.17=h166bdaf_0
liblapack=3.9.0=19_linux64_openblas
libopenblas=0.3.24=pthreads_h413a1c8_0
libpng=1.6.39=h753d276_0
libsqlite=3.43.2=h2797004_0
libstdcxx-ng=13.2.0=h7e041cc_2
libtiff=4.5.1=h6a678d5_0
libuuid=1.41.5=h5eee18b_0
libwebp=1.3.2=h11a3e52_0
libwebp-base=1.3.2=hd590300_0
libxcb=1.16=hd590300_0
libxml2=2.9.14=h22db469_4
libzlib=1.2.13=hd590300_5
markdown=3.5=pypi_0
markupsafe=2.1.2=pypi_0
matplotlib-base=3.4.3=py39h2fa2bec_2
matplotlib-inline=0.1.6=pypi_0
mistune=3.0.2=pypi_0
mpmath=1.3.0=pypi_0
multidict=6.0.4=pypi_0
multiprocess=0.70.15=pypi_0
nbclassic=0.5.6=pypi_0
nbclient=0.8.0=pypi_0
nbconvert=7.9.2=pypi_0
nbformat=5.9.2=pypi_0
ncurses=6.4=hcb278e6_0
nest-asyncio=1.5.8=pypi_0
networkx=3.0=pypi_0
notebook=6.5.6=pypi_0
notebook-shim=0.2.3=pypi_0
numpy=1.24.1=pypi_0
oauthlib=3.2.2=pypi_0
openjpeg=2.5.0=hfec8fc6_2
openssl=1.1.1w=hd590300_0
packaging=23.2=pypi_0
pandas=2.1.2=py39hddac248_0
pandocfilters=1.5.0=pypi_0
parso=0.8.3=pypi_0
pcre=8.45=h9c3ff4c_0
pexpect=4.8.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=9.3.0=pypi_0
pip=23.3.1=pyhd8ed1ab_0
pixman=0.42.2=h59595ed_0
platformdirs=3.11.0=pypi_0
prometheus-client=0.17.1=pypi_0
prompt-toolkit=3.0.39=pypi_0
protobuf=4.23.4=pypi_0
psutil=5.9.6=pypi_0
pthread-stubs=0.4=h36c2ea0_1001
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.2=pypi_0
pyarrow=13.0.0=pypi_0
pyasn1=0.5.0=pypi_0
pyasn1-modules=0.3.0=pypi_0
pycairo=1.23.0=py39hd1222b9_0
pycparser=2.21=pypi_0
pyg-lib=0.3.0+pt21cu118=pypi_0
pygments=2.16.1=pypi_0
pyparsing=3.1.1=pyhd8ed1ab_0
python=3.9.0=hffdb5ce_5_cpython
python-dateutil=2.8.2=pyhd8ed1ab_0
python-tzdata=2023.3=pyhd8ed1ab_0
python_abi=3.9=4_cp39
pytorch-fast-transformers=0.4.0=pypi_0
pytorch-lightning=1.1.5=pypi_0
pytz=2023.3.post1=pyhd8ed1ab_0
pyyaml=6.0.1=pypi_0
pyzmq=24.0.1=pypi_0
rdkit=2022.03.2=py39h89e00b9_1
readline=8.2=h8228510_1
referencing=0.30.2=pypi_0
regex=2023.10.3=pypi_0
reportlab=3.6.12=py39h5eee18b_0
requests=2.31.0=pypi_0
requests-oauthlib=1.3.1=pypi_0
rpds-py=0.10.6=pypi_0
rsa=4.9=pypi_0
s3transfer=0.7.0=pypi_0
sacremoses=0.0.53=pypi_0
scikit-learn=1.3.2=pypi_0
scipy=1.11.3=pypi_0
send2trash=1.8.2=pypi_0
sentencepiece=0.1.99=pypi_0
setuptools=68.2.2=pyhd8ed1ab_0
six=1.16.0=pyh6c4a22f_0
sniffio=1.3.0=pypi_0
soupsieve=2.5=pypi_0
sqlalchemy=2.0.22=py39hd1e30aa_0
sqlite=3.43.2=h2c6b66d_0
stack-data=0.6.3=pypi_0
sympy=1.12=pypi_0
tensorboard=2.15.0=pypi_0
tensorboard-data-server=0.7.2=pypi_0
terminado=0.17.1=pypi_0
threadpoolctl=3.2.0=pypi_0
tinycss2=1.2.1=pypi_0
tk=8.6.13=h2797004_0
tokenizers=0.10.3=pypi_0
torch=2.1.0+cu118=pypi_0
torch-geometric=2.4.0=pypi_0
torch-scatter=2.1.2+pt21cu118=pypi_0
torch-sparse=0.6.18+pt21cu118=pypi_0
torchaudio=2.1.0+cu118=pypi_0
torchvision=0.16.0+cu118=pypi_0
tornado=6.3.3=py39hd1e30aa_1
tqdm=4.49.0=pypi_0
traitlets=5.12.0=pypi_0
transformers=4.6.0=pypi_0
triton=2.1.0=pypi_0
typing-extensions=4.4.0=pypi_0
typing_extensions=4.8.0=pyha770c72_0
tzdata=2023c=h71feb2d_0
urllib3=1.26.13=pypi_0
wcwidth=0.2.8=pypi_0
webencodings=0.5.1=pypi_0
websocket-client=1.6.4=pypi_0
werkzeug=3.0.1=pypi_0
wheel=0.41.2=pyhd8ed1ab_0
widgetsnbextension=3.6.6=pypi_0
xorg-libxau=1.0.11=hd590300_0
xorg-libxdmcp=1.1.3=h7f98852_0
xxhash=3.4.1=pypi_0
xz=5.4.2=h5eee18b_0
yarl=1.9.2=pypi_0
zipp=3.17.0=pypi_0
zlib=1.2.13=hd590300_5
zstd=1.5.5=hfc55251_0
